requirements.txt
PyYAML>=6.0
colorama>=0.4.6
PyQt5>=5.15.0
pyinstaller>=6.0.0
